requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.4.0=pypi_0
aiofiles=22.1.0=pypi_0
aiosqlite=0.18.0=pypi_0
anyio=3.6.2=pypi_0
appnope=0.1.3=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.2.3=pypi_0
asttokens=2.2.1=pypi_0
astunparse=1.6.3=pypi_0
attrs=22.2.0=pypi_0
babel=2.12.1=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.0=pypi_0
bleach=6.0.0=pypi_0
bzip2=1.0.8=h3422bc3_4
c-ares=1.18.1=h3422bc3_0
ca-certificates=2022.12.7=h4653dfc_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.3.0=pypi_0
certifi=2022.12.7=pypi_0
cffi=1.15.1=pypi_0
charset-normalizer=3.1.0=pypi_0
comm=0.1.3=pypi_0
contourpy=1.0.7=pypi_0
cycler=0.11.0=pypi_0
debugpy=1.6.6=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
et-xmlfile=1.1.0=pypi_0
executing=1.2.0=pypi_0
fastjsonschema=2.16.3=pypi_0
flatbuffers=23.3.3=pypi_0
fonttools=4.39.3=pypi_0
fqdn=1.5.1=pypi_0
gast=0.4.0=pypi_0
google-auth=2.17.1=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.53.0=pypi_0
h5py=3.6.0=nompi_py38hacf61ce_100
hdf5=1.12.1=nompi_hd9dbc9e_104
idna=3.4=pypi_0
importlib-metadata=6.1.0=pypi_0
importlib-resources=5.12.0=pypi_0
ipykernel=6.22.0=pypi_0
ipython=8.12.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.6=pypi_0
isoduration=20.11.0=pypi_0
jax=0.4.8=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
json5=0.9.11=pypi_0
jsonpointer=2.3=pypi_0
jsonschema=4.17.3=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.1.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.3.0=pypi_0
jupyter-events=0.6.3=pypi_0
jupyter-server=2.5.0=pypi_0
jupyter-server-fileid=0.8.0=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyter-server-ydoc=0.8.0=pypi_0
jupyter-ydoc=0.2.3=pypi_0
jupyterlab=3.6.3=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.22.0=pypi_0
jupyterlab-widgets=3.0.7=pypi_0
keras=2.12.0=pypi_0
kiwisolver=1.4.4=pypi_0
krb5=1.20.1=h69eda48_0
libblas=3.9.0=16_osxarm64_openblas
libcblas=3.9.0=16_osxarm64_openblas
libclang=16.0.0=pypi_0
libcurl=7.88.1=h9049daf_1
libcxx=16.0.0=h75e25f2_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=12_2_0_hd922786_31
libgfortran5=12.2.0=h0eea778_31
liblapack=3.9.0=16_osxarm64_openblas
libnghttp2=1.52.0=hae82a92_0
libopenblas=0.3.21=openmp_hc731615_3
libprotobuf=3.19.4=hccf11d3_0
libsqlite=3.40.0=h76d750c_0
libssh2=1.10.0=h7a5bd25_3
libzlib=1.2.13=h03a7124_4
llvm-openmp=16.0.1=h7cfbb63_0
markdown=3.4.3=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=2.0.5=pypi_0
ml-dtypes=0.0.4=pypi_0
nbclassic=0.5.4=pypi_0
nbclient=0.7.2=pypi_0
nbconvert=7.2.10=pypi_0
nbformat=5.8.0=pypi_0
ncurses=6.3=h07bb92c_1
nest-asyncio=1.5.6=pypi_0
notebook=6.5.3=pypi_0
notebook-shim=0.2.2=pypi_0
numpy=1.23.2=py38h579d673_0
oauthlib=3.2.2=pypi_0
openpyxl=3.1.2=pypi_0
openssl=3.1.0=h03a7124_0
opt-einsum=3.3.0=pypi_0
packaging=23.0=pypi_0
pandas=1.5.3=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
patsy=0.5.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.5.0=pypi_0
pip=23.0.1=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=3.2.0=pypi_0
prometheus-client=0.16.0=pypi_0
prompt-toolkit=3.0.38=pypi_0
protobuf=4.22.1=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pypi_0
pygments=2.14.0=pypi_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.8.16=h3ba56d0_1_cpython
python-dateutil=2.8.2=pypi_0
python-json-logger=2.0.7=pypi_0
python_abi=3.8=3_cp38
pytz=2023.3=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.0.2=pypi_0
qtconsole=5.4.1=pypi_0
qtpy=2.3.1=pypi_0
readline=8.2=h92ec313_1
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=pypi_0
send2trash=1.8.0=pypi_0
setuptools=67.6.1=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pypi_0
soupsieve=2.4=pypi_0
stack-data=0.6.2=pypi_0
statsmodels=0.13.5=pypi_0
tensorboard=2.12.1=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow-deps=2.10.0=0
tensorflow-estimator=2.12.0=pypi_0
tensorflow-macos=2.12.0=pypi_0
tensorflow-metal=0.8.0=pypi_0
termcolor=2.2.0=pypi_0
terminado=0.17.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=he1e0b03_0
tomli=2.0.1=pypi_0
tornado=6.2=pypi_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pypi_0
typing-extensions=4.5.0=pypi_0
uri-template=1.2.0=pypi_0
urllib3=1.26.15=pypi_0
wcwidth=0.2.6=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.5.1=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.40.0=pyhd8ed1ab_0
widgetsnbextension=4.0.7=pypi_0
wrapt=1.14.1=pypi_0
xz=5.2.6=h57fd34a_0
y-py=0.5.9=pypi_0
ypy-websocket=0.8.2=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h03a7124_4
zstd=1.5.2=hf913c23_6
